+++ CHANGELOG.md
@@ -9,6 +9,18 @@ Fixed       for any bug fixes.
 Security    to invite users to upgrade in case of vulnerabilities.
 -->
 
+## 2.4.0 - 2023/10/16
+
+### Added
+
+- prompt to continue excluding incompatible flags
+- e2e testing for `install` and `add` commands
+
+### Fixed
+
+- `global` flag position on `yarn` and `yarn@berry`
+- `package` parameter on `add` command
+
 ## 2.3.1 - 2023/10/04
 
 ### Fixed